% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandeR.R
\name{addScoresToNewShapefile}
\alias{addScoresToNewShapefile}
\title{Augment an existing shapefile by adding scores}
\usage{
addScoresToNewShapefile(in_filename, out_filename, scores = c("all"))
}
\arguments{
\item{in_filename}{Filename of the shapefile to be scored.}

\item{out_filename}{Filename of the shapefile that will hold the results.}

\item{scores}{List of scores to include, a subset of `getListOfScores()`}
}
\value{
The filename of the new shapefile.
}
\description{
Since analysis often happens at a high level, after scores have been
       shapefile itself. This function does this.
}
\examples{
library(mandeR)
newname <- tempfile()
mandeR::addScoresToNewShapefile(mass_cd(), newname)

}
