% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{AreaCalc}
\alias{AreaCalc}
\title{Calculate the area for one or multiple polygons in a
       Shapefile}
\usage{
AreaCalc(x, y, id = NA)
}
\arguments{
\item{x}{X-coordinates of the polygon nodes}

\item{y}{Y-coordinates of the polygon nodes}

\item{id}{ids which indicate to which polygons each xy-coordinate belongs}
}
\value{
The area of polygon(s)
}
\description{
Returns a vector of values with the area for all 
       polygons in the Spatial Data Frame.
}
\examples{
  
library(sp)
library(ggplot2)
library(plyr)
library(gerry)
#identify the path of the sample shapefile of congressional districts
shapefilepath     <- mass_cd()                    
#create a spatial data frame
shape             <- readOGR(shapefilepath)       
#Specify and equal area coordinate reference system
proj              <- test<-CRS("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 
                                +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 
                                +datum=NAD83 +units=m")
#reproject the data to the specified CRS
reproj            <- spTransform(shape,proj)      
#Create an ID number for each polgyon
reproj@data["id"] <- rownames(shape@data)                   
#create a vector of the nodes in the polygons for the spatial data frame
shpvector         <- fortify(reproj)              
sdict             <- shpvector \%>\% filter(group==0.1)
#Calculate the area of the polygons and return a list of values
parea        <- AreaCalc(sdict$long, sdict$lat, sdict$id) 
#Add a column of the perimeter values to the spatial data frame
shpvector@data["Area"]<- parea         
ggplot()+
  geom_polygon(data=shape3,aes(x=long,y=lat,group=group,
               fill=shpvector$Area))     
 
}
