% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{getListOfScores}
\alias{getListOfScores}
\title{Retrieve a list of available scores}
\usage{
getListOfScores()
}
\value{
A list of scores names, which are represented by character strings.
}
\description{
mandeR, through compactnesslib, has access to a growing number of
       compactness scores. This function provides a list of them. Subsets of
       this list can be passed to other functions to restrict which scores
       are calculated. Descriptions of scores may be found in the vignettes.
}
