% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{getScoresForGeoJSON}
\alias{getScoresForGeoJSON}
\title{Get a dataframe of scores from a GeoJSON input}
\usage{
getScoresForGeoJSON(geojson_str, id = "", scores = c())
}
\arguments{
\item{geojson_str}{GeoJSON representing features to be scored.}

\item{id}{Attribute id to key features in the resulting table. If
omitted, numeric keys are returned based on the input
order of the features.}

\item{scores}{List of scores to include, a subset of `getListOfScores()`}
}
\value{
A data frame with scores as columns, including the id column.
}
\description{
GeoJSON is used to pass information between compactnesslib and its
       higher level interfaces. (Well-known text could be used, but we don't
       have a parser for that, yet. Volunteer?) This function takes a GeoJSON
       description of one or more multipolygons and returns their scores as a
       data table.
}
\examples{
library(mandeR)
dists <- sf::st_read(mass_cd())
gj    <- geojsonio::geojson_json(dists)
mandeR::getScoresForGeoJSON(gj, 'DIST_NUM')

}
