% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandeR.R
\name{getScoresForWKT}
\alias{getScoresForWKT}
\title{Get a dataframe of scores from a WKT input}
\usage{
getScoresForWKT(wkt_str, scores = c("all"))
}
\arguments{
\item{wkt_str}{WKT representing feature to be scored.}

\item{scores}{List of scores to include, a subset of `getListOfScores()`}
}
\value{
A data frame with scores as columns, including the id column.
}
\description{
WKT is used to pass information between compactnesslib and its
       higher level interfaces. This function takes a WKT string of a polygon
       or a multipolygons and returns its scores as a data frame. The
       coordinates must be appropriate: latlong cannot be used.
}
\examples{
library(mandeR)
dists   <- sf::st_read(mass_cd())
wkt_str <- st_as_text(st_geometry(dists)[[1]])
mandeR::getScoresForWKT(wkt_str)

}
