% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{mass_cd}
\alias{mass_cd}
\title{Get path to dggrid executable}
\usage{
mass_cd()
}
\value{
A string representing the path to the dggrid executable.
}
\description{
Returns a path to the dggrid executable. Used for running stuff.
}
