% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{PolsbyPopper}
\alias{PolsbyPopper}
\title{Calculate Polsby Popper Metric for one or multiple polygons in a
       Shapefile}
\usage{
PolsbyPopper(nodevector)
}
\arguments{
\item{x}{X-coordinates of the polygon nodes}

\item{y}{Y-coordinates of the polygon nodes}
}
\value{
The Polsby Popper metrics
}
\description{
Returns a vector of values with the Polsby Popper metric for all 
       polygons in the Spatial Data Frame.
}
\examples{
  
library(sp)
library(ggplot2)
library(plyr)
library(gerry)
#identify the path of the sample shapefile of congressional districts
shapefilepath     <- mass_cd()                    
#create a spatial data frame
shape             <- readOGR(shapefilepath)       
#Specify and equal area coordinate reference system
proj              <- test<-CRS("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 
                                +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 
                                +datum=NAD83 +units=m")
#reproject the data to the specified CRS
reproj            <- spTransform(shape,proj)      
#Create an ID number for each polgyon
reproj@data["id"] <- rownames(shape@data)                   
#create a vector of the nodes in the polygons for the spatial data frame
shpvector         <- fortify(reproj)              
sdict             <- shpvector \%>\% filter(group==0.1)
#Calculate the perimeter of the polygons and return a list of values
PolPop        <- PolsbyPopper(sdict$long, sdict$lat, sdict$id) 
#Add a column of the perimeter values to the spatial data frame
shpvector@data["PolsbyPopper"]<- PolPop         
ggplot()+
  geom_polygon(data=shape3,aes(x=long,y=lat,group=group,
               fill=shpvector$PolsbyPopper))     
 
}
