% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandeR.R
\name{augmentShapefileWithScores}
\alias{augmentShapefileWithScores}
\title{Augment an existing shapefile by adding scores}
\usage{
augmentShapefileWithScores(filename, scores = c("all"))
}
\arguments{
\item{filename}{Filename of the shapefile to be altered.}

\item{scores}{List of scores to include, a subset of `getListOfScores()`}
}
\value{
The filename of the modified shapefile.
}
\description{
Since analysis often happens at a high level, after scores have been
       shapefile itself. This function does this. The shapefile's projection
       must be appropriate: latlong cannot be used.
}
\examples{
\dontrun{
library(mandeR)
mandeR::augmentShapefileWithScores("path/to/myshapefile.shp", c("Reock", "area"))
}

}
