% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandeR.R
\name{mass_cd}
\alias{mass_cd}
\title{Get an example shapefile}
\usage{
mass_cd()
}
\value{
A string representing the path to the example shapefile
}
\description{
Returns the path of an example shapefile of the electoral districts of 
       Massachusetts.
}
