% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{PerimeterCalc}
\alias{PerimeterCalc}
\title{Calculate the perimeter of all polygons in a spatial data frame}
\usage{
PerimeterCalc(nodevector)
}
\arguments{
\item{nodevector}{A vector of the XY coordinates for all nodes in a polygon - generated using sp.fortify()}
}
\value{
A vector of the perimeter measurements for each polygon in the Spatial Data Frame.
}
\description{
Returns a vector of values with the perimeters for all polygons in the Spatial Data Frame.
}
\examples{

   
library(sp)
library(ggplot2)
library(plyr)
shapefilepath     <- mass_cd()                    #identify the path of the sample shapefile of congressional districts
shape             <- readOGR(shapefilepath)       #create a spatial data frame
#Specify and equal area coordinate reference system
proj              <- test<-CRS("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m")
reproj            <- spTransform(shape,proj)      #reproject the data to the specified CRS
reproj@data["id"] <- rownames(shape@data)         #Create an ID number for each polgyon
shape2            <- fortify(reproj)              #fortify the data, converting it to a vector of XY values
shpvector         <- fortify(reproj)              #create a vector of the nodes in the polygons for the spatial data frame
shpperim          <- PerimeterCalc(shpvector)     #Calculate the perimeter of the polygons and return a list of values
shpvector@data["Perimeter"]<-shpperim             #Add a column of the perimeter values to the spatial data frame
ggplot()+
  geom_polygon(data=shape3,aes(x=long,y=lat,group=group,fill=shpvector$Perimeter))     
 
}
